backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.26.59
botocore==1.29.165
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dataclasses-json==0.6.6
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.2
django-allauth==0.53.0
django-cors-headers==4.2.0
django-filter==24.2
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dj-rest-auth==3.0.0
drf-yasg==1.21.7
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.23
langchain-openai==0.0.5
langsmith==0.0.87
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.31.0
openapi-codec==1.3.2
packaging==23.2
pillow==10.3.0
psycopg2-binary==2.9.6
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.6.2
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
stripe==10.1.0
tenacity==8.3.0
tiktoken==0.5.2
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
yarl==1.9.4
